create home page|add icons to nav bar
