+++ src/routes/Home.tsx
@@ -1,252 +1,112 @@
 import {
-  AppShell,
-  Blockquote,
-  Burger,
-  Flex,
-  Paper,
-  Card,
-  Image,
-  Text,
-  Badge,
   Button,
-  Group,
-  Timeline,
-  Accordion,
+  Card,
   Divider,
-  Title,
+  Paper,
+  Text,
+  Image,
+  Center,
+  Flex,
 } from "@mantine/core";
-import { useDisclosure } from "@mantine/hooks";
+import { IconArrowRight, IconPhoto } from "@tabler/icons-react";
+import React from "react";
+//import { Button } from '../components/Button'
+import { useNavigate } from "react-router";
 
-import { useState } from "react";
-
-import { BarChart } from "@mantine/charts";
-import {
-  IconGitBranch,
-  IconGitPullRequest,
-  IconGitCommit,
-  IconMessageDots,
-} from "@tabler/icons-react";
-
-const data = [
-  { language: "Python", years: 5 },
-  { language: "JavaScript", years: 4 },
-  { language: "Java", years: 3 },
-  { language: "C++", years: 2 },
-  { language: "Go", years: 1 },
-  { language: "Ruby", years: 1 },
-];
-const groceries = [
-  {
-    emoji: "📚",
-    value: "Education",
-    description: (
-      <Timeline active={3} bulletSize={24} lineWidth={2}>
-        <Timeline.Item bullet={<IconGitBranch size={12} />} title="High School">
-          <Text c="dimmed" size="sm">
-            Cape Town High
-          </Text>
-          <Text size="xs" mt={4}>
-            2016-2020
-          </Text>
-        </Timeline.Item>
-
-        <Timeline.Item bullet={<IconGitCommit size={12} />} title="University">
-          <Text c="dimmed" size="sm">
-            Cape Peninsula University of Technology
-          </Text>
-          <Text size="xs" mt={4}>
-            2021-2025
-          </Text>
-        </Timeline.Item>
-
-        <Timeline.Item
-          bullet={<IconGitCommit size={12} />}
-          title="Super University"
-        >
-          <Text c="dimmed" size="sm">
-            Udemy
-          </Text>
-          <Text size="xs" mt={4}>
-            2025 - present
-          </Text>
-        </Timeline.Item>
-      </Timeline>
-    ),
-  },
-  {
-    emoji: "💻",
-    value: "Work Experience",
-    description: (
-      <Timeline active={1} bulletSize={24} lineWidth={2}>
-        <Timeline.Item
-          bullet={<IconGitBranch size={12} />}
-          title="Free Lancing"
-        >
-          <Text c="dimmed" size="sm">
-            Java Developer
-          </Text>
-          <Text size="xs" mt={4}>
-            2024-2025
-          </Text>
-        </Timeline.Item>
-
-        <Timeline.Item
-          bullet={<IconGitCommit size={12} />}
-          title="Devloper Intern"
-        >
-          <Text c="dimmed" size="sm">
-            Full Stack intern
-          </Text>
-          <Text size="xs" mt={4}>
-            2025-2026
-          </Text>
-        </Timeline.Item>
-
-        <Timeline.Item
-          bullet={<IconGitCommit size={12} />}
-          title="Full Stack Developer"
-        >
-          <Text c="dimmed" size="sm">
-            Junior Developer
-          </Text>
-          <Text size="xs" mt={4}>
-            2026 -
-          </Text>
-        </Timeline.Item>
-      </Timeline>
-    ),
-  },
-];
-type HomeProps = {
-  setMain: (main: string) => void;
-};
-const Home: React.FC<HomeProps> = ({ setMain }) => {
+const Home = ({ setMain }: { setMain: (main: string) => void }) => {
   return (
-    <Card shadow="sm" radius="md" className="outer-container">
-      <Card
-        shadow="sm"
-        padding="lg"
-        radius="md"
-        withBorder
+    <div
+      style={{
+        height: "calc(100vh - 40px)",
+        width: "calc(100vw - 412px)",
+        backgroundColor: "rgba(75, 136, 197, 0.998)",
+        background: "linear-gradient(125deg,lightblue ,white )",
+      }}
+    >
+      <div
         style={{
-          width: "600px",
+          height: "100%",
+          width: "100%",
           display: "flex",
-          flexDirection: "column",
-          justifyContent: "space-between",
+          alignItems: "center",
+          justifyContent: "center",
         }}
       >
-        <div>
-          <Card.Section>
-            <Image src="/profile.jpg" height={450} alt="profile picture" />
-          </Card.Section>
-
-          <Group justify="space-between" mt="md" mb="xs">
-            <Text fw={500}>Profile</Text>
-            <Badge variant="light" color="pink">
-              Open to New Roles
-            </Badge>
-          </Group>
-
-          <div
-            style={{
-              display: "flex",
-              flexDirection: "row",
-              justifyContent: "space-between",
-            }}
-          >
-            <Text>Dillyn Lakey</Text>
-            <Text size="sm" c="dimmed">
-              084 2654 522
-            </Text>
-          </div>
-          <Group>
-            <Text>dillyn32lakey23@gmail.com</Text>
-          </Group>
-        </div>
-        <div>
-          <Blockquote
-            style={{ height: "90%", marginTop: "10px", marginBottom: "10px" }}
-          >
-            {/* <Paper shadow="md" p="xl" style={{ marginBottom: "15px" }}>A */}
-            A passionate developer skilled in building sleek UIs with React.js
-            and scalable backends with Node.js. Expert in JavaScript/TypeScript,
-            focused on clean code and solving problems efficiently. A team
-            player who thrives on challenges and keeps up with the latest tech
-            trends.
-            {/* </Paper> */}
-          </Blockquote>
-        </div>
-
-        <div
-          style={{ width: "100%", display: "flex", justifyContent: "center" }}
+        <Card
+          shadow="1px 0px 27px 0px rgba(0,0,0,0.2)"
+          p="xl"
+          radius="xl"
+          className="main-container"
         >
-          <Button
-            onClick={() => setMain("/profile")}
-            variant="outline"
-            size="lg"
-            radius="xl"
-          >
-            View full profile
-          </Button>
-        </div>
-      </Card>
-      <div className="second-inner-container">
-        <Card className="first-second-inner-container">
-          <div
-            style={{
-              display: "flex",
-              alignItems: "center",
-              flexDirection: "column",
-            }}
-          >
-            <Text>Programming language experience in years</Text>
-            <BarChart
-              yAxisLabel="Years"
-              xAxisLabel="Languages"
-              h={300}
-              data={data}
-              dataKey="language" // X-axis labels
-              series={[{ name: "years", color: "blue" }]}
-              styles={{
-                axis: { fontSize: "20px" }, // Axis labels
-                tooltip: { fontSize: "12px" },
+          <div className="main-container-column">
+            <div
+              style={{
+                paddingTop: "100px",
+                paddingLeft: "50px",
+                lineHeight: "2.8",
               }}
-            />
-          </div>
-        </Card>
-        <Card className="second-second-inner-container">
-          <Title size={"16px"}>Work Experience:</Title>
-          <Text>
-            As a freelance developer, I’ve delivered custom web and mobile
-            applications, specializing in both frontend and backend development.
-            I’ve worked on various projects, including e-commerce platforms and
-            content management systems, using technologies like React, Node.js,
-            and PHP to create responsive, user-friendly solutions.
-          </Text>
-          <Title size={"16px"} style={{ marginTop: "10px" }}>
-            Education:
-          </Title>
-          <Text>
-            I’m currently studying Application Development at Cape Peninsula
-            University, where I’ve gained a solid foundation in programming,
-            software development, and building efficient applications.
-          </Text>
+            >
+              <Text className="" h="64px" size="68px" fw={900} c={"black"}>
+                I Build Custom
+              </Text>
+              <Text
+                h="64px"
+                size="68px"
+                fw={900}
+                variant="gradient"
+                gradient={{ from: "black", to: "black", deg: 360 }}
+              >
+                Web Solutions
+              </Text>
+              <Text h="64px" size="68px" fw={900} c={"black"}>
+                That Help Your
+              </Text>
+              <Text
+                span
+                h="70px"
+                size="80px"
+                fw={900}
+                variant="gradient"
+                gradient={{ from: "blue", to: "indigo", deg: 90 }}
+              >
+                Business Grow
+              </Text>
+              <Text
+                c={"black"}
+                size="16px"
+                fw={450}
+                style={{ paddingTop: "15px" }}
+              >
+                I help you increase conversion and performance in every possible
+                way.
+              </Text>
 
-          <Divider style={{ marginTop: "25px" }} />
-          <Accordion style={{ overflowY: "scroll" }}>
-            {groceries.map((item) => (
-              <Accordion.Item key={item.value} value={item.value}>
-                <Accordion.Control icon={item.emoji}>
-                  {item.value}
-                </Accordion.Control>
-                <Accordion.Panel>{item.description}</Accordion.Panel>
-              </Accordion.Item>
-            ))}
-          </Accordion>
+              <Button
+                w={"500px"}
+                variant="filled"
+                rightSection={<IconArrowRight size={32} />}
+                onClick={() => setMain("/details")}
+                size="xl"
+                radius="md"
+                style={{ marginTop: "25px" }}
+              >
+                View Details Page
+              </Button>
+            </div>
+            <div>
+              <Image
+                radius={400}
+                height={800}
+                w={640}
+                src="/profile.jpg"
+                alt="profile picture"
+              />
+            </div>
+          </div>
         </Card>
       </div>
-    </Card>
+    </div>
   );
-}
+};
 
-export default Home
+export default Home;
