--- src/routes/Profile.tsx
@@ -1,18 +1,0 @@
-
-import { Button, Card } from '@mantine/core';
-import React from 'react'
-//import { Button } from '../components/Button'
-import { useNavigate } from 'react-router'
-
-const Profile = () => {
-  return (
-    <Card shadow="sm" radius="md" className="outer-container">
-      Profile
-    
-    </Card>
-  
-  )
-}
-
-
-export default Profile
